requirements.txt
Django==2.1.3
django-environ==0.4.5
djangorestframework==3.9.0
# psycopg2==2.7.6
psycopg2-binary==2.7.6
colorama==0.4.0
